[PhpUnitBridge] Modernize CoverageListener
